- refactor transaction
